remove refine_qualities
the idea is to move to longest-edge
preference, where length is measured
using the size field.
I suspect this will do just about the
same thing on regular meshes,
and hopefully do well enough on
real meshes in combination with shape
correction.
Git will forever preserve the code
if we want it again