more flexible cycle prevention
in the 3D tests, we are seeing that a high
quality floor for coarsening results in
bad elements being repetitively refined
and the resulting elements are only worse
so they still can't be coarsened.
we can prevent cycles and also allow
collapses in very bad cavities
by using the quality of the mesh
before sliver splitting as the quality floor.
the sliver splits and subsequent coarsen
should really be treated as a compound
operation, and it is their compound application
which should not be allowed to decrease quality.