indset chains can push very bad splits
if we're using the mark-all-edges
sliver method, then several adjacent
slivers can cause a chain reaction in
the indset logic that will continuously
accept really bad splits.
the solution for now will be to not give
the sliver refinement a zero floor, rather
a small but nonzero floor.
this was discovered when doing coarsening
before refinement, switched away from that now.