specify malloc/realloc behavior with zero size
note that this affects the checks we just put
in place, i.e. a zero pointer is now returned
for something which is "allocated", but that
is a rare enough case that I'd rather not
introduce a bunch of variables just to detect
it. those are asserts, not production logic.