fix owners_from_global using comm_graph_exact
we rely on the "inverse graph communicator" (gci)
to put things back exactly in the order that
the forward communicator transmitted them, so we
have to explicitly specify the incoming messages
and their offsets.
before, we were letting MPI_Dist_graph_create
derive them, and it doesn't come up with the
original layout (nor should it).