ensure that malloc calls don't return NULL
this allows us to use a convention in the rest
of the code that a NULL array pointer means
that array has not been allocated, where
previously it could also mean the array
was allocated with size zero.
pointers are now non-zero regardless of the
amount of memory they point to.