silence covered switch default in Clang
this is a controversial issue.
*if* the program is operating as expecting,
an enum variable will only have one of the enum
values and thus having a default case is pointless.
Clang is smart enough to see this.
we were avoiding default cases, but the GCC
and NVCC compilers were complaining.
so now we have default cases, so on the off chance
that an enum is not in range, the failure mechanics
are a bit clearer ?
they really should just abort or something...