"positivize" the RIB axis
an axis and its negative create
the same bisection, but the sign
defines which half of the ranks
get which half of the points.
this makes an arbitrary but consistent
decision between such axes, which
is helpful to get the desired
mapping of partitions to rank
IDs, though not necessarily
optimal.