RIB: perturb axis in case of co-planar points
very much a hack, but also shows up often
in practice when people use highly structured
meshes.
we roughly detect a set of coplanar points
near the center as an inability to bisect
with less than 5% imbalance, and slightly
perturb the components of the axis in hopes
of escaping that special-case plane.
simple example: four-triangle mesh in 2D.