remove assert from bridge_graph
this check was valid before we
started supporting bridge_dual_graph.
it is meant to assert that the input
is an undirected graph converted to
a directed graph, hence the number
of directed edges must be even.

however, the dual graph has directed
edges to a dummy node, and the dummy
node has a huge index (2^32-1) and
doesn't return those directed edges,
which breaks this check.
the code is carefully crafted to
not ask too many questions about
this dummy node.

this assert started getting hit
when we started deriving 2D edges
via the triangle dual graph.