adding man pages to Git
these are generated by the 'doctext'
tool which is part of the 'sowing'
package, used by PETSc / MPICH
but not very standard.

I worry that people will simply
not go through the hassle of installing
sowing and as a consequence will not
get their man pages.

They are small enough text that although
they are auto-generated documents,
I'm willing to add them to Git to ensure
that everyone who gets the code
has the man pages for the API.