use constexpr for MPI_Datatype static init
From Stack Overflow:
"Prior to C++11, only static const data members of integral or
enumeration type could have initializers in the class definition."
The reason this worked before is we were compiling with MPICH,
and they define MPI_Datatype as an integer type, so it was legal
C++98. OpenMPI defines MPI_Datatype as a pointer to a struct,
so we can only do this static init thing in C++11 with constexpr.